add a new componet popupbox. edit inquiry row componet to displat only topic. edit farmerinquiry , dealerInquiry files. CRUD working (update issue-not display)
