+++ frontend/src/pages/Veenath/farmerInquiry.js
@@ -1,11 +1,13 @@
 import React, { useEffect, useState } from 'react';
 import axios from 'axios';
 import InquiryRow from '../../Component/Veenath/InquiryRow';
-import { Link } from 'react-router-dom';
+import InquiryDetailsPopup from '../../Component/Veenath/InquiryDetailsPopup';
 
 const FarmerInquiry = () => {
   const [pendingInquiries, setPendingInquiries] = useState([]);
   const [resolvedFarmerInquiries, setResolvedFarmerInquiries] = useState([]);
+  const [selectedInquiry, setSelectedInquiry] = useState(null);
+  const [isPopupOpen, setIsPopupOpen] = useState(false);
 
   useEffect(() => {
     const fetchInquiries = async () => {
@@ -23,14 +25,14 @@ const FarmerInquiry = () => {
     fetchInquiries();
   }, []);
 
-  const handleDelete = async (id) => {
-    try {
-      await axios.delete(`http://localhost:8070/api/reports/${id}`);
-      setPendingInquiries(pendingInquiries.filter(inquiry => inquiry._id !== id));
-      setResolvedFarmerInquiries(resolvedFarmerInquiries.filter(inquiry => inquiry._id !== id));
-    } catch (error) {
-      console.error(error);
-    }
+  const handleViewInquiry = (inquiry) => {
+    setSelectedInquiry(inquiry);
+    setIsPopupOpen(true);
+  };
+
+  const handleClosePopup = () => {
+    setSelectedInquiry(null);
+    setIsPopupOpen(false);
   };
 
   return (
@@ -40,11 +42,8 @@ const FarmerInquiry = () => {
         <h2>Pending Inquiries</h2>
         {pendingInquiries.map((inquiry) => (
           <div key={inquiry._id}>
-            <InquiryRow inquiry={inquiry} />
-            <button onClick={() => handleDelete(inquiry._id)}>Delete</button>
-            <Link to={`/formPage?id=${inquiry._id}`}>
-              <button>Update</button>
-            </Link>
+            <div>Topic: {inquiry.topic}</div>
+            <button onClick={() => handleViewInquiry(inquiry)}>View</button>
           </div>
         ))}
       </div>
@@ -52,11 +51,18 @@ const FarmerInquiry = () => {
         <h2>Resolved Inquiries</h2>
         {resolvedFarmerInquiries.map((inquiry) => (
           <div key={inquiry._id}>
-            <InquiryRow inquiry={inquiry} />
-            <button onClick={() => handleDelete(inquiry._id)}>Delete</button>
+            <div>Topic: {inquiry.topic}</div>
+            <button onClick={() => handleViewInquiry(inquiry)}>View</button>
           </div>
         ))}
       </div>
+      {isPopupOpen && (
+        <div className="popup-background">
+          <div className="popup">
+            <InquiryDetailsPopup inquiry={selectedInquiry} onClosePopup={handleClosePopup} />
+          </div>
+        </div>
+      )}
     </div>
   );
 }
